FA archiver: Fix a subtle bug involving FA ids
The computation of the array of ids to retrieve involves some simple
integer arithmetic ... and it turns out that matlab has some unpleasant
surprises here.  In brief: int32(5)/10 == 1 .  Ah.

So, make sure we're working with doubles (fewer surprises here), both when
assembling the request mask in fa_load, and in the result returned by
fa_name2id.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@129797 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ matlab/fa_load.m
@@ -401,7 +401,7 @@ function result = format_mask(mask, max_id)
 
     % Assemble array of ints from ids and send as raw mask array
     mask_array = zeros(1, max_id / 32);
-    for id = mask
+    for id = double(mask)
         ix = floor(id / 32) + 1;
         mask_array(ix) = bitor(mask_array(ix), bitshift(1, mod(id, 32)));
     end
